update: move widgets to view/
