+++ lib/view/chat/chat_bubble.dart
