+++ lib/view/chat/chat_list.dart
@@ -1,5 +1,5 @@
 import 'package:chatapp/viewmodel/chat_view_model.dart';
-import 'package:chatapp/widget/chat_bubble.dart';
+import 'package:chatapp/view/chat/chat_bubble.dart';
 import 'package:flutter/material.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
