+++ lib/view/chat/chat_page.dart
@@ -1,6 +1,6 @@
 import 'package:chatapp/view/chat/chat_list.dart';
 import 'package:chatapp/viewmodel/chat_view_model.dart';
-import 'package:chatapp/widget/chat_text_field.dart';
+import 'package:chatapp/view/chat/chat_text_field.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_hooks/flutter_hooks.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
